[BEAM-1397] Move metrics related tests to MetricsTest
